bugfix: incorrect nullmap handling on bulk-exec mode
In case when bulk-exec mode, it may fetch a column with not-null constraint
being already on tcache. This tcache_column_store does not has null-map,
so we fill-up kernel buffer using clEnqueueFillBuffer().
It incorrectly used get_global_offset(0), so some bytes were not initialized.

In addition, kern_column_to_column() which is used to load a tcache columnar-
data into in-kernel sorting chunk (that is larger column-store).
Its NULL-bitmap handing was not correct; what we had to do is set a bit
(1 << (index & 31)), but what we did is (1 << (index & ~31))!.