fixbug: gpuscan exited relation scan unintentionally.
NOTE: In case when it scans on a large hole (that is
continuous blocks contain invisible tuples only; may
be created by DELETE with relaxed condition),
pgstrom_data_store_insert_block() may return negative
value without valid tuples, even though we don't reach
either end of relation or chunk.
So, we need to check whether we actually touched on
the end-of-relation. If not, retry scanning.