bugfix: codegen.c generated multiple functions with same name
PostgreSQL allows function overload; a function with same name
but different argument types shall be considered as a different
function, but OpenCL (based on C99) does not allow overloading.
So, we need qualifiers of function name, if it may have multiple
variations.

It was reported by Anzai Naoya, at the bug #97