bugfix: gpupreagg didn't rewrite the plan tree correctly.
Agg node assumes the input stream comes from OUTER_VAR, thus expression
node is also set-up to reference variable in ecxt_outertuple.
On the other hands, because of custom-scan interface contract, expression
nodes in GpuPreAgg node has to reference variables in ecxt_innertuple
(that is aliasn of INDEX_VAR or other usual varno).
So, we had to fixup varno of expression on the GpuPreAgg node by ourselves
because GpuPreAgg works after the setrefs.c logic.

This bug was reported by Anzai Naoya.