bugfix: CpuReCheck on GpuPreAgg referenced broken krowmap
Old style bulkload moved data store and row-map pair, however, it
became obsolete. So, it should not be used any more.
On the other hands, new GpuPreAgg uses kern_row_map to store the
index to the partial aggregation, and it is written back to the
host side with DMA.
It means, krowmap (if exists) shall be broken by write-back DMA,
and it may have non-negative nvalids; that leads reference to
the rowmap even if it is invalid.