bugfix: incorrect wipe out of junk target-entry
TupleDesc of ss_ScanTupleSlot in ScanState is initialized according
to the custom_ps_tlist, if not associated with a particular relation.
On ExecInitCustomScan, ExecCleanTypeFromTL() makes the TupleDesc,
however, it wipes out target-entry with resjunk=true.
If we intend to reference the column, we have to keep resjunk=true.
(Elsewhere, this pseudo entry is used to only for EXPLAIN output)