fixbug: gpujoin_join_quals() was not __synchthreads() aware
In GpuNestLoop logic, gpujoin_join_quals() has to be called by
all the GPU threads even though the thread is mapped on the
out-of-range items, because it internally synchronize local
threads to reduce DRAM accesses.
So, this commit adjust specification of gpujoin_join_quals()
to handle NULL input on outer_buffer and inner_htup.