GpuPreAgg is applied even though it will not work effectively
In case when number of groups are equal or close to number to tuples
per chunk, GpuPreAgg will not work effectively, or make more waste
of time.
At this moment, we did cost estimation more carefully, however,
it may be able to improve the coverage if we expand buffer size
until row reduction makes sense.