bugfix: incorrect PlaceHolderVar handling
When GpuJoin takes underlying subquery, we have see PlaceHolderVar
to reference expression that is already computed in the sub-query.
(v9.5 specific) GpuProjection may look at an expression on the
tlist_dev, however, we have to treat it as if var reference.
Thus, IsA(node, Var) is not reliable check no longer. Instead of
the previous manner, we have to check ps_src_depth/ps_src_resno.
If TLE is Var/PHV, it has valid index, not -1.