bugfix: GpuJoin can skip a virtual partition window, if resize failed on second major retry.
gpujoin_resize_window() may give up next trial with smaller virtual
partition size if kds_dst is already consumed 50% of the buffer.
In this case, GpuJoin kernel does not return error code because
the join results of earlier half are successfully processed.
However, a pair of window_base and window_size now points the range
where GpuJoin does not process yet, so we had to revert these
variables prior to the exit of kernel function.