bugfix: inner table of GpuJoin should not be larger than 4GB.
Because of the current data format contract of KDS_FORMAT_ROW/HASH,
row_index points HeapTuple using 32bit offset. It means we cannot
have KDS with KDS_FORMAT_ROW/HASH larger than 4GB.
Only inner tables of GpuJoin can have such a large KDS chunk.

It should be enhanced to 32GB in the future version.
The least 3bit is not used because all the items are aligned to 64bit.