cuda_modules_lock is protected by mutex, not spinlock
When spinlock wait is blocked for a long time, PostgreSQL makes
crash the process, so it should be used for time consuming tasks.
On the other hands, NVRTC needs time to build a code if it is
very complicated.
It implies we should not protect a portion to build GPU code by
spinlocks.