bugfix: ProjectionPath prevents combined GpuJoin by host-side projection
MEMO: See grouping_planner() where it calls create_projection_path()
on the partial Path-nodes. It forcibly injects ProjectionPath to
have individual PathTarget on top of the scan/join paths, because
these scan/join paths may be referenced by other path-trees, thus
it is unable to set PathTarget in-place.
In fact, create_projection_plan() pulls up sub-plan and attach its
target-list if it can be compatible, however, this ProjectionPath
prevents GpuJoin to have compatible host_tlist and dev_tlist, and
leads host-side projection. It works as a blocker of combined-GpuJoin
which is one of the performance key of reporting queries.
So, we duplicate GpuJoinPath by ourself, and set PathTarget of
the ProjectionPath here.