add dedicated CUDA C program builder background worker
It looks to us (suspicious at least) NVRTC of CUDA 9.1 has a problem
when nvrtcCompileProgram() is kicked from worker thread.
Once this API is called from the worker thread, proc_exit callback
handler calls something in libnvrtc, then it leads segmentation fault.
Even though we are not 100% certain, we stop to use worker thread of
GpuContext as kicker of nvrtcCompileProgram().
Instead of the old implementation, we launch a background worker for
asynchronous code build which does not use any worker threads.

(gdb) bt
   from /usr/local/cuda-9.1/targets/x86_64-linux/lib/libnvrtc.so.9.1
   from /usr/local/cuda-9.1/targets/x86_64-linux/lib/libnvrtc.so.9.1
    listp=0x7efd132ef6c8 <__exit_funcs>,
    run_list_atexit=run_list_atexit@entry=true) at exit.c:77
    argv=argv@entry=0x12ec000, dbname=0x12ebf50 "nvme",
    username=<optimized out>) at postgres.c:4298
    argv=argv@entry=0x12a39e0) at postmaster.c:1363