correct wrong abstraction to scan table
GpuJoin/GpuPreAgg may call GpuScan logic when it pulls up outer
table scan, however, it is not an adequate abstraction.
This commit moves the portion to scan an outer table into common
gpu_task.c logic, and GpuTaskState had a common shared-state to
control table scan.
It also helps to adopt a common logic for parallel scan because
ParallelHeapScanDesc was re-defined at PG11, thus, it makes us
to put complicated #if ... #endif logic.