pg_numeric_t was re-defined to use Int128_t for internal representation.
This update is a groundwork for Decimal type in Apache Arrow.
It has fixed-point 128bit integer format, and we plan to map it on numeric
of PostgreSQL. We now no longer use numeric data type for atomic operations,
so here is no strong reason why pg_numeric_t is mapped to 64bit integer using
our own data format.