workaround: discount of inner cost of GpuJoin as if parallel-scan
FIXME: This is an ad-hoc workaround not to breal multi-level
GpuJoin chain. PostgreSQL v11 begins to support parallel build
of inner hash-table, and its cost tends to be discounted by
num of parallel workers. It is exactly a rapid option to process
inner side, however, cannot use combined GpuJoin+GpuPreAgg here.
So, we tentatively discount the inner cost by parallel_nworkers,
as if it is processed in parallel.
It shall be supported on the later PG-Strom version.