add pg_strom.regression_test_mode to disable some EXPLAIN output
If PGDATA locates on NVME volume, PG-Strom automatically find out
the closest GPU and EXPLAIN shows which is the preferable one.
However, it is specific to the environment where you want to run
the regression test.
pg_strom.regression_test_mode can be used to disable these kind
of problematic output.