add pg_strom.reuse_cuda_context
allocation and destruction of CUDA context takes about 125ms in avg.
it is unignorable extra time consumption for short queries.
this option allows to preserve CUDA context after the query execution,
and reuse it.

however, note that we have to create CUDA context when we run parallel
workers. So, it makes sense only when a single backend process runs
small query (e.g, GPU memory store)