bugfix: compiler unintentionally wipe out resource/lock cleanup at STROM_CATCH()
From the compiler point of view, we may have no code path that set gstore_mapped
between STROM_CATCH and STROM_END_TRY. However, an exception might be caught by
the STROM_TRY(setsigjmp) then it branches to the STROM_CATCH block after the
gstore_mapped is turned on.
In this case, a compiler optimization for invocation of gstoreFdwUnmapDeviceMemory
makes resource/lock leaks. To avoid the problem, we add volatile qualifier on
the gstore_mapped variable.

{
    bool   gstore_mapped = false;

    STROM_TRY();
    {
        if (pds_src->kds.format == KDS_FORMAT_COLUMN)
        {
            rc = gstoreFdwMapDeviceMemory(GpuWorkerCurrentContext, pds_src);
            if (rc != CUDA_SUCCESS)
                werror("failed on gstoreFdwMapDeviceMemory: %s", errorText(rc));
            gstore_mapped = true;
        }
        retval = __gpuscan_process_task(gtask, cuda_module);
    }
    STROM_CATCH();
    {
        if (gstore_mapped)
            gstoreFdwUnmapDeviceMemory(GpuWorkerCurrentContext, pds_src);
        STROM_RE_THROW();
    }
    STROM_END_TRY();
    if (gstore_mapped)
        gstoreFdwUnmapDeviceMemory(GpuWorkerCurrentContext, pds_src);
}