bugfix: Gstore_Fdw potentially has corrupted PK-index
PK-index entry must be created after the allocation of row-id, and
PK-index entry must be remoted before the release of row-id.
The prior implementation had a short timing when PK-index entry
points a row-id already released.
This commit force to create Undo log of 'A' (add PK) next to the
'I' (insert), and 'R' (remove PK) prior to 'D' (delete), to ensure
row-id is occupied during PK-index is valid.