arrow_fdw: fixup alignment checks on the length of RecordBatchFieldState
Even though Apache Arrow requires values array must be aligned to
64bit (hopefully, 512bit for AVX-512 operations), it never defines
length of the values array.
So, we removed the check whether nullmap/values/extra_length is
power of 64bit, or not.
Even though values_length does not fit 64bit boundary, the next
chunk shall start from the 64bit aligned address, so here is no
problems. Even if the target chunk is the tail of Apache Arrow
file, it always has footer area. So, it shall never overrun.

This problem was reported by Eric Lu.