divergenced threads must be converged manually, for safety.
Volta or later GPU architecture has PC register for each thread.
It allows to run a part of threads in a warp execute individual
instruction once branch operations divergented threads.
A set of simultaneous threads are represented by __activemask().
Once if (...) branched a part of threads, it divergents from
others, then it prevents to use __shfl_sync() or others.
So, if logic uses warp level datum exchange and synchronization,
we have to pay attention to inject __syncwarp(~0U) to converge
the threads after the operations that lead thread divergence.