gpu_cache: avoid concurrent reference to old-version of varlena datum
GPU Cache reuses old-version of varlena datum if it is not changed
on UPDATE, however, the older code could reference uninitialized
pointer if INSERT old-tuple and UPDATE new-tuple are applied by
the same GPU kernel concurrently.
This commit changed the policy a bit, to reuse the old varlena only
if the older version is already committed on the GPU cache side.
(It means the older version is never modified by the kernel.)