bugfix: __threadfence() ensures updates of global hash visible
gpupreagg_global_reduction() once allocates and initializes
a slot on kds_final and hash-item, then updates the hash-chain
by atomic-operations.
Even if hash-chain is updated, values on the kds_final and f_hash
may not visible to other threads because CUDA adopts weakly-
ordered memory model.
So, we added __threadfench() prior to atomicExch().