bugfix: flat-buffer default should not be non-zero
We had a small optimization when scalar values are added to
flat-buffer if it is identical with the "default" value.
However, non-zero default value is not reliable because it
assumes the reader code can reproduce the default value.
When ArrowType::Time was expanded to have bitWidth, we
"incorrectly" assigned its default value to 32, but it leads
unexpected metadata corruption at the reader side.