gpupreagg: kds_final's tuple use L1 cache line exclusively
In case when multiple tuples on the kds_final are stored on one
L1 cache line, the formerly written one can be loaded to L1 cache
on the other SM, without later written tuple.
Then, it breaks memory access consistency when we reference the
later written tuples.

Right now, it is not certain whether it is GPU's problem or
compiler's problem. However, we can avoid this problem by L1 cache
like alignment, so it writes out the grouping tuples according to
the L1 cache line size alignment.