remove: custom_path_remember() infrastructure
It was reported at #618.
When # of relations involved in a join exceeds the threshold of GEQO,
PostgreSQL switches optimization logic from the default one.
GEOQ internally uses its own memory context, thus, plan-nodes are
eventually released at end of the GEOQ optimization routine.
It makes impossible to track unused path-nodes, so we changed the
logic to merge joins with multiple-relations.