bugfix: wrong order of kds_final projection leads misaligned access
This commit fixes the issue reported at #636 and #637.

In the older version, aggregate function appeared only in HAVING
clause was ordered after the grouping-keys. It has no guarantee
the intermediate state is aligned to 64bit. Then, following
atomic operation might cause misaligned access.

Also, if grouping-key is the primary key, it means individual
tuples shall not be grouped. Thus, PostgreSQL allows to use
the columns that are not listed as grouping-keys.