disallow to compute a part of expression at the projection
The issue #717 reported CASE ... WHEN expression was not pushed down
into the device execution, and part of expressions are distributed
on the tlist_dev. It leads unexpected error like 'division by zero'
that should be avoided by the given condition.
Fundamentally, it is not possible to determine which part of the
expression is guarded, thus we give up to calculate a part of
expression, and adopts all-or-nothing rule on the projection.