bugfix: assertion failure at cost_memoize_rescan()
If Path->rows would be a non-integer value, it may cause assertion
failure at cost_memoize_rescan() due to the calculation logic of
hit_ratio.
It is now discussed in pgsql-hackers, and will be fixed at upstream,
however, this change will avoid problem on the older versions.

issue related to #794