CREATE OPERATOR makes a pseudo commutor/negator operators in the default namespace
issue at #805
When we create '=' operator, its NEGATOR operator should be '<>'
however, it may not be created yet. In this case, PostgreSQL
create a pseudo operator in the default namespace (usually, it
is in 'public' namespace). Next, when we create the '<>' operator
in the 'pg_catalog', it violates circuler reference checks.