bugfix: a bunch of rows were skipped after GPU kernel suspend/resume
issue #819

When GPU kernel raised CPU-fallback first time, it once suspends
the GPU kernel because the fallback buffer is not attached on.
Then, it restarts GPU kernel, and it fetched rows according to
the wp->smx_row_count.
In the old implementation, we incremented the wp->smx_row_count
immediately on the early half of __gpuscan_load_source_xxxx().
It means, if CPU-fallback raised GPU-kernel suspend/resume,
we will see the next block even though the suspended block was
not processed.